+++ mean_square_displacement/include/msd/common.hpp
