+++ mean_square_displacement/include/msd/debugTools.cuh
