+++ mean_square_displacement/include/msd/defines.h
@@ -1,0 +1,13 @@
+#ifndef MSDCONFIG_H
+#define MSDCONFIG_H
+#define MeanSquareDisplacement_VERSION_MAJOR 0
+#define MeanSquareDisplacement_VERSION_MINOR 0
+/* #undef USE_BOOST */
+/* #undef USE_CUDA */
+/* #undef USE_CPU */
+
+
+#ifdef USE_CUDA
+#define USE_GPU
+#endif
+#endif
