+++ mean_square_displacement/include/msd/gitversion.h
@@ -1,0 +1,6 @@
+#ifndef GITVERSION_H
+#define GITVERSION_H
+#define GITVERSION "b58543c7d082d4408343d26f9e7b420790b60a05"
+
+#define MSD_VERSION "0.0.0"
+#endif
