+++ mean_square_displacement/include/msd/gitversion.h.in
