+++ mean_square_displacement/include/msd/kahan.hpp
